server: move makeSQLServerArgs
This function is used by the tenant code but is in `testserver.go`. I
tend to ignore `testserver.go` when I'm trying to understand a
non-testing code path so this was confusing.

This commit moves the function to `tenant.go` and gives it a more
specific name.

Release note: None
